+++ src/core/core.mjs
@@ -1,0 +1,3 @@
+export const getJSONValue = (jsonObj, key) => {
+  return key.split('.').reduce((obj, prop) => obj?.[prop], jsonObj)
+}