+++ test.config.js
@@ -1,0 +1,5 @@
+const config = {
+  testDir: 'test',
+}
+
+export default config