+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- getValue() function
+- unit tests
+
 ## 1.0.1 - 2024-10-17
 ### Added
 - Project initialization